+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added target argument for command line. `m3u-filter -t <target_name> -t <target_name>`. Target names should be provided in the config.
 * Added filter to mapper definition.
 * Refactored filter parsing.
+* Fixed sort after mapping group names.
 
 # v0.9.9(2023-03-20)
 * Added optional 'enabled' property to input and target. Default is true.